==> PreProcessing/1_category_feature_engineering.ipynb <==
# Feature engineering on Item feature categories.


Item features can hold many interesting information, but there is a very high amount of them.

Some item share some feature categories (such as color, size) but some don't.

Different features can be engineered from item featues, such as:

* The most common visited (feature_category, feature_value) pairs. ✅
* The number of visits on the most common pair. ✅

import os 
import warnings
warnings.filterwarnings('ignore')
import pandas as pd

# launch this cell if you have issues on windows with py4j (think about updating your PATH)
import sys
os.environ['PYSPARK_DRIVER_PYTHON_OPTS']= "notebook"
os.environ['PYSPARK_DRIVER_PYTHON'] = sys.executable
os.environ['PYSPARK_PYTHON'] = sys.executable

# starts a spark session from notebook

os.environ['PYSPARK_SUBMIT_ARGS'] ="--conf spark.driver.memory=4g  pyspark-shell"
from pyspark.sql import SparkSession

spark = SparkSession \
    .builder \
    .master("local[*]") \
    .appName("load_explore") \
    .getOrCreate()

# loads relevant datas in DataFrames
train_sessions = spark.read.load('../Data/train_sessions.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

train_purchases = spark.read.load('../Data/train_purchases.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

candidate_items = spark.read.load('../Data/candidate_items.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

item_features = spark.read.load('../Data/item_features.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')


mapped_item_features = item_features.rdd.map(lambda x: (x.item_id, (x.feature_category_id, x.feature_value_id)))

item_session_and_features = (train_sessions.rdd
                             .map(lambda x: (x.item_id, (x.session_id)))  # Maps sessions to a (key, value pair)
                             .join(mapped_item_features)  # Huge operation, might be too heavy
)

item_session_and_features.take(5)

from collections import Counter
# Groupping by session too heavy, let's try to perform a reduce operation

# At this point, item_features_per_session contains tuples of (item_id, (sessions_id, (feature_cat_id, feature_val_id))
# Let's reorder the tuple to (session_id, (item_id, feature_cat_id, feature_val_id))

def reduce_count_item_pairs(x, y):
    '''Merges the y dictionnary into the x dictionnary.
    Expects to have both x and y as dictionnary having the keys (feature_cat_id, feature_val_id)
    and a counter of occurences as a value.
    '''
    for feature_pair in y:
        if x.get(feature_pair) is None:
            x[feature_pair] = y[feature_pair]
        else:
            x[feature_pair] += y[feature_pair]
    return x
        
item_features_per_session = (item_session_and_features
                             .map(lambda x: (x[1][0], (x[0], x[1][1][0], x[1][1][1])))  # Reorder to (session_id, (item_id, feature_cat_id, feature_val_id))
                             .mapValues(lambda x: {(x[1], x[2]): 1})  # Transforms to (session_id, counter_of(feature_cat_id, feature_val_id))
                             .reduceByKey(reduce_count_item_pairs)  # Reduce by merging the dictionaries and their values
                             .mapValues(lambda x: Counter(x).most_common()[0])  # Takes the most common key tuple
                             .mapValues(lambda x: (x[0][0], x[0][1], x[1]))
)

item_features_per_session.take(5)

## Counting number of categories.

We want to provide features engineered from the item feature categories. There are many feature categories. Let's count them.

def reduce_unique(x, y):
    ''' Merges all the sets togheter '''
    return (x[0], x[1] | y[1])

unique_categories = (item_features.rdd
                    .map(lambda x: (x.item_id, (x.feature_category_id)))  # Maps every item feature entry to (item_id, (feature_category_id))
                    .mapValues(lambda x: set([x]))  # Maps the values to a set, so they will be counted as unique during the reduce opetations.
                    .reduce(reduce_unique)  # Reduces the values by merging the sets, eliminating duplicate feature_category_ids
)

len(unique_categories[1])

There are 73 different feature categories. 

## Engineering our first 73 features

Some features could be engineered such as counting the number of occurences for each category met per session.
We will have 73 different features only for that.

def count_categories(x, y):
    ''' Merges all the dictionnary values togheter, adding the categories values alltogheter '''
    for key in y:
        x[key] = x.get(key, 0) + y[key]
    return x

def fill_empties(x, nb_categories=len(unique_categories[1])):
    ''' Fill the dictionnary with categories that are not present
    setting their occurence count to 0 '''
    for idx in range(0, nb_categories+1):
        x[idx] = x.get(idx, 0)
    return x

def get_occurences(x):
    ''' Remaps the values so we only have the occurences. 
    The feature category id by itself is defined by the order (from 1 to 73).
    '''
    listed_x = list(x)
    new_list = []
    for elem in listed_x:
        new_list.append(elem[1])
    return tuple(new_list)

feature_categories_per_session = (item_session_and_features
                             .map(lambda x: (x[1][0], (x[1][1][0])))  # Project to (session_id, (feature_cat_id))
                             .mapValues(lambda x: {x: 1})  # Maps value before being reduced, putting them in a dictionnary
                             .reduceByKey(count_categories)  # Counts feature categories occurence in a dictionnary
                             .mapValues(fill_empties)  # Fills unmet categories and sets their occurence to 0
                             .mapValues(lambda x: tuple(sorted(x.items(), key=lambda x: x[0])))  # Sorts by key value
                             .mapValues(get_occurences)
)

feature_categories_per_session.take(1)

## Feature groupping using frequent itemset

The dimensionality can be reduced. For example, we merge the feature categories inside groups, determined from frequent itemset analysis.

Spark ML package has Frequent Pattern Mining functionalities, implementing FP-Growth for example.  

def occurences_to_present_category(x):
    ''' Maps a list of occurences (where the feature category id is defined by the order)
    to a list containing the feautre categories with a non null occurency.
    '''
    present_categories = set()
    assert(len(x) == 74) # Checks that the data has 73 features
    for idx in range(74):
        if x[idx] != 0:
            present_categories.add(idx)
    return tuple(present_categories)
            

present_categories_per_session = feature_categories_per_session.mapValues(occurences_to_present_category)

present_categories_per_session.take(1)

present_categories_itemsets = present_categories_per_session.map(lambda x: list(x[1])).cache()  # Dropping the sessions data

from pyspark.mllib.fpm import FPGrowth

# Starts training the FPGrowth tree on the pipelined RDD
model = FPGrowth.train(present_categories_per_session.map(lambda x: list(x[1])).cache(), minSupport=0.4, numPartitions=10)

# model.freqItemsets().count()

# Selecting the biggest itemsets

Now that we have itemsets with a support of at least 20%, let's select a number of large groups and use them as features.

If in a session, the categories visited at least once are equal or a superset of one of those groups, sets the value to 1; else 0.

# Compute the largest itemsets and sorts them by the length of the groups (largest groups first)
frequent_itemsets = model.freqItemsets().sortBy(lambda x: len(x.items), ascending=False, numPartitions=4).cache()

# top_20_frequent = frequent_itemsets.take(20)

# top_20_frequent

We have a problem: the most frequent itemsets found here have a big part in common. The ideal would be to find frequent itemsets but with fairly different items.

# Item feature category clustering

One technique would be to perform clustering on item category features. For each session, we take the feature occurences per session, and perform normalization.

An arbitrary number of clusters X could be created. From that, we could one-hot encode the classified cluster of each session and create from that X new features in our dataset.

import numpy as np

def normalize_vector(x):
    values = np.array(x, dtype=np.float64)
    norm = np.linalg.norm(values)
    values[:] = values[:] / norm
    return values

normalized_categories = feature_categories_per_session.mapValues(normalize_vector).map(lambda x: x[1])

from pyspark.mllib.clustering import KMeans, KMeansModel

number_of_clusters = 10

# Build the model (cluster the data)
clusters = KMeans.train(normalized_categories, number_of_clusters, maxIterations=10, initializationMode="random")

# Evaluate clustering by computing Within Set Sum of Squared Errors
def error(point):
    center = clusters.centers[clusters.predict(point)]
    return np.sqrt(sum([x**2 for x in (point - center)]))

WSSSE = normalized_categories.map(lambda point: error(point)).reduce(lambda x, y: x + y)

print("Within Set Sum of Squared Error = " + str(WSSSE))

# Set seed to draw always the same vector
np.random.seed(42)

# We try our clusters by creating a fake vector and predicting its cluster.
category_vector = np.random.normal(loc=10.0, scale=5.0, size=len(unique_categories[1])+1)

# We clip the negative values to 0 
negative_values = category_vector < 0.0
category_vector[negative_values] = 0.0

category_vector

# We predict our vector values
clusters.predict(category_vector)

# Feature engineering on item feature cateogries

Now that we have our clusters, we perform a simple mapValues operation to cluster each item session and to one hot encode the results.

def one_hot_encode_integer(x, max_range=number_of_clusters):
    values = np.zeros(max_range, dtype=np.uint8)
    values[x-1] = 1
    return values
    

clustered_sessions = feature_categories_per_session.mapValues(normalize_vector).mapValues(lambda x: clusters.predict(x)).mapValues(one_hot_encode_integer)

clustered_sessions.take(10)
