==> PreProcessing/1_features_engineering.ipynb <==
# Data preprocessing and Feature Engineering

In order to obtain a learnable dataset, features must be preprocessed and engineered in order to contain valuable learning data. This notebook will generate a PySpark RDD that contains the preprocessed dataset, ready for feature selection algorithms.

More than 20 features must be engineered, some ideas are:

Session-time related features:
* Weekday of the session ❌
* Month of the session ✅
* Season of the session ✅
* Duration of the session ✅
* Day period of the session ✅
* Is it during the weekend? ❌

Session-item related features:
* Number of viewed items per session ✅
* Number of unique viewed items per session ✅
* Item with the most time spent on. ✅
* Longest time spend on item in the session. ✅
* The last visited item of the session ✅
* The first visited item on the session ✅
* Mean time per item. ✅
* Variance time per item. ✅
* The most revisited item. ✅
* How much time this item has been revisited. ✅
* The amount of items revisited more than once. ✅

### Starting the Spark engine and loading the dataset.

import os 
import warnings
warnings.filterwarnings('ignore')
import pandas as pd

# launch this cell if you have issues on windows with py4j (think about updating your PATH)
import sys
os.environ['PYSPARK_DRIVER_PYTHON_OPTS']= "notebook"
os.environ['PYSPARK_DRIVER_PYTHON'] = sys.executable
os.environ['PYSPARK_PYTHON'] = sys.executable

# starts a spark session from notebook

os.environ['PYSPARK_SUBMIT_ARGS'] ="--conf spark.driver.memory=4g  pyspark-shell"
from pyspark.sql import SparkSession

spark = SparkSession \
    .builder \
    .master("local[*]") \
    .appName("load_explore") \
    .getOrCreate()

# loads relevant datas in DataFrames
train_sessions = spark.read.load('../Data/train_sessions.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

train_purchases = spark.read.load('../Data/train_purchases.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

candidate_items = spark.read.load('../Data/candidate_items.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

item_features = spark.read.load('../Data/item_features.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

datas = [train_sessions, train_purchases, candidate_items, item_features]

## Group sessions and aggregate their duration.

Using a map reduce operation, we will start by computing the session duration in seconds. For that, we use a function that will convert the timestamp into seconds since the 1st January 1970 (UNIX time).

import datetime

def timestamp_to_unix(timestamp: str) -> int:
    '''Converts the timestamp, on a string format to the UNIX time
    '''
    try:
        date = datetime.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S.%f')
    except ValueError:
        date = datetime.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S')
    return int(date.timestamp())

posix_date_mapped_sessions = train_sessions.rdd.map(lambda x: (x.session_id, timestamp_to_unix(x.date))).cache()
sampled_sessions = posix_date_mapped_sessions.sample(False, 0.0001, seed=42)

sampled_sessions.take(1)

def reduce_min_max(x, y):
    if type(x) == tuple:
        return(max(x[0], y), min(x[1], y))
    return (max(x, y), min(x, y))

reduced_sessions = posix_date_mapped_sessions.reduceByKey(lambda x, y: reduce_min_max(x, y))
reduced_sessions.take(1)

def get_duration_seconds(tupl):
    if type(tupl[1]) is int:
        return (tupl[0], 0)
    else:
        return (tupl[0], tupl[1][0] - tupl[1][1])

time_sessions = reduced_sessions.map(lambda x: get_duration_seconds(x))

time_sessions.take(10)

We cannot compute the sessions that only show one item, so we set the session duration to 0

## Group sessions and aggregate the number of unique AND total viewed items

We compute how much unique and total items have been viewed during a session

item_mapped_sessions = train_sessions.rdd.map(lambda x: (x.session_id, x.item_id)).cache()
uniquely_viewed_items = item_mapped_sessions.groupByKey().mapValues(lambda vals: len(set(vals)))

uniquely_viewed_items.take(10)

total_viewed_items = item_mapped_sessions.groupByKey().mapValues(lambda vals: len(list(vals)))
total_viewed_items.take(10)

# Time period sessions evaluation

We will map for each session its day in the week, if it is on weekend, the month, the season and the year.

# Computing the starting date for each session
starting_date_sessions = train_sessions.rdd.map(lambda x: (x.session_id, x.date)).reduceByKey(lambda x, y: min(x, y))

type(starting_date_sessions.take(1)[0])

def parse_datetime(timestamp):
    try:
        return datetime.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S.%f')
    except ValueError:
        return datetime.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S')
    

def get_season(date_time):
    season = (date_time.month - 1) // 3
    season += (date_time.month == 3)&(date_time.day>=20)
    season += (date_time.month == 6)&(date_time.day>=21)
    season += (date_time.month == 9)&(date_time.day>=23)
    season -= 3*int(((date_time.month == 12)&(date_time.day>=21)))
    return season

def get_day_period(date_time):
    '''Converts the date_time into the day of the week.
    
    0 -> Morning (from 6am to 12am)
    1 -> Afternoon (from 12am to 6pm)
    2 -> Evening (from 6pm to 12pm)
    3 -> Night (from 12pm to 6am)
    '''
    return date_time.hour // 4
    
def map_day_period(timestamp):
    # Converts into datetime
    date_time = parse_datetime(timestamp)
    # Computes the season
    season = get_season(date_time)
    # Computes the 
    return (get_day_period(date_time), date_time.weekday(), int(date_time.weekday() > 4), date_time.month - 1, season, date_time.year)

date_period_sessions = starting_date_sessions.map(lambda x: (x[0], map_day_period(x[1])))
date_period_sessions.take(10)

# Time per item evaluation

Computing the items where time was the most spent on. The time is considered to be the one between two clicks.

As it is impossible to determine how much time was spent on the last item of the session, the information will be computed from the train_purchase dataset, where the date of the purchase is also included.

def compute_item_duration(item_list):
    '''Return a new list of tuples, where the first element is the item_id and the second element
    is its related duration. The duration is computed from the time difference with the next visited item.
    
    :param item_list: A list containing tuples of (item_id, datetime)
    '''
    new_list = []
    for idx in range(len(item_list)):
        if idx != len(item_list) - 1:
            time_begin = parse_datetime(item_list[idx][1])
            time_end = parse_datetime(item_list[idx+1][1])
            delta_seconds = (time_end - time_begin).seconds
            new_list.append((item_list[idx][0], delta_seconds))
        else:
            new_list.append((item_list[idx][0], item_list[idx][1]))
    return new_list
            
    


time_per_item = (train_sessions.rdd
    .map(lambda x: (x.session_id, (x.item_id, x.date)))  # For every entry maps a tuple with key=session_id and values=(item_id, date)
    .groupByKey()  # Groups by session_id, result (key: session_id, values: Iterable[item_id, date])
    .mapValues(lambda x: sorted(list(x), key=lambda _x: _x[1]))  # For each key, maps the values to a sorted version
    .mapValues(lambda x: compute_item_duration(x)))  # For each value: Iterable[item_id, date] maps to a new Iterable[item_id, duration]



for it in time_per_item.take(1)[0][1]:
    print(it)

def compute_last_item_time(item_list, purchase_time):
    '''Computes the time spent on the last item.
    :param item_list: A list of visited items in a session, already processed by the compute_item_duration function
    :param purchase_time: The timestamp of the last purchased item
    '''
    new_list = []
    for idx in range(len(item_list)):
        if idx != len(item_list) - 1:
            new_list.append(item_list[idx])
        else:
            begin_time = parse_datetime(item_list[idx][1])
            end_time = parse_datetime(purchase_time)
            delta_seconds = (end_time - begin_time).seconds
            new_list.append((item_list[idx][0], delta_seconds))
    return new_list


time_per_item = (train_purchases.rdd
    .map(lambda x: (x.session_id, x.date))  # For every entry maps a tuple with key=session_id and value=x.date
    .join(time_per_item)  #  Joins the purchase dataset with the time_per_item dataset from the key (session_id)
    .mapValues(lambda x: compute_last_item_time(x[1], x[0]))  # Sets the duration of the last viewed item.
)

time_per_item.take(10)

# First feature and second features: item with the most time spent on and the spent time.
longest_items = time_per_item.mapValues(lambda x: sorted(x, key=lambda _x : _x[1])[-1][0])
longest_items.take(10)

time_per_item.take(10)

# Third feature: the last item visited in the session.
def sort_items_by_date(x):
    '''Sorts the items from the first visited to the last visited.
    :param x: An iterable of (item_id, date) tuples.
    '''
    values = [(item, parse_datetime(date)) for item, date in x]
    values = sorted(values, key=lambda x: x[1])
    return values


first_and_last_item_per_session = (train_sessions.rdd.map(lambda x: (x.session_id, (x.item_id, x.date)))
                                 .groupByKey()
                                 .mapValues(sort_items_by_date)
                                 .mapValues(lambda x: (x[0][0], x[-1][0]))
)

first_and_last_item_per_session.take(10)

import numpy as np

# Fourth and fifth features: the mean and standard deviation of the time spend on each item 
mean_time_per_session = time_per_item.mapValues(lambda x: (np.array(x)[:, 1].mean(), np.array(x)[:, 1].std()))
mean_time_per_session.take(10)

def count_different_items(x):
    '''This function will count for each session the revisited item, 
    returning for each session a list of tuples (item_id, revisits) 
    
    :param x: A list of item visits on each session.
    '''
    values = [x_ for x_ in x] 
    values_dict = {}
    for item_id, _ in values:
        if values_dict.get(item_id) is None:
            values_dict[item_id] = 1
        else:
            values_dict[item_id] += 1
    return sorted([(k, v) for k, v in values_dict.items()], key=lambda x: x[1], reverse=True)

def most_revisited_items(x, default_value=-1):
    '''This function, usually concatenated to most_revisited_items,
    returns the item that has been the most revisited with the number of revisits.
    As well as the amount of items that have been revisited more than once.
    
    The input items are considered to be sorted by the number of revisits.
    
    If no item was revisited more than once, it will return (default_value, default_value, 0)
    '''
    if x[0][1] <= 1:  # Items visited at most 0 times
        return (default_value, default_value, 0)
    else:
        # Count the number of items that have been revisited at least once.
        counter = 0
        for item_id, revisits in x:
            if revisits > 1:
                counter += 1
            else:
                break  # As we consider that the list is sorted by the number of revisits.

        return (x[0][0], x[0][1], counter)
    

items_per_sessions = (train_sessions.rdd
                      .map(lambda x: (x.session_id, (x.item_id, 1)))
                      .groupByKey()
                      .mapValues(count_different_items)
                      .mapValues(most_revisited_items)
)

items_per_sessions.take(10)
