==> PreProcessing/3_KNN.ipynb <==
import os 
import warnings
import time
import numpy as np
import scipy.stats
import sys
import sklearn
import sklearn.datasets

from pyspark.sql import SparkSession
warnings.filterwarnings('ignore')
import pandas as pd

# launch this cell if you have issues on windows with py4j (think about updating your PATH)

os.environ['PYSPARK_DRIVER_PYTHON_OPTS']= "notebook"
os.environ['PYSPARK_DRIVER_PYTHON'] = sys.executable
os.environ['PYSPARK_PYTHON'] = sys.executable

# starts a spark session from notebook

os.environ['PYSPARK_SUBMIT_ARGS'] ="--conf spark.driver.memory=4g  pyspark-shell"


spark = SparkSession \
    .builder \
    .master("local[*]") \
    .appName("feature_selection") \
    .getOrCreate()

sc=spark.sparkContext

train_sessions_engineered = spark.read.load('../Data/session_engineered_features.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true').cache()
columns_df = train_sessions_engineered.columns

train_purchases = spark.read.load('../Data/train_purchases.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')


#### Data Normalization ([0,...,1])

summary = train_sessions_engineered.describe().toPandas().set_index('summary').transpose()
#print(summary)
columns_to_normalize = ['session_time','most_time_spent_on_item','number_o_visit','mean_time','number_o_revisited_items','std_time']+[str(i) for i in range(1,25)]+['_45']
dic_val = {}
for elem in columns_to_normalize:
    index = np.where(np.array(train_sessions_engineered.columns) == elem)[0][0]
    row = summary.iloc[index]
    dic_val[index] = (float(row[3]),float(row[4]))
print(dic_val)

broadcast_clusters = sc.broadcast(dic_val)
def normalize_column(row, br):
    dic = br.value
    dickeys = dic.keys()
    temp = []
    for i in range(45):
        if i in dic:
            elem = dic.get(i)
            temp.append((row[i]-elem[0])/elem[1])
        else:
            temp.append(row[i])
    return tuple(temp)
print(normalize_column(train_sessions_engineered.take(1)[0], broadcast_clusters))
train_sessions_engineered = train_sessions_engineered.rdd.map(lambda x: normalize_column(x,broadcast_clusters)).toDF(columns_df)

#### Keep only useful columns

mrmr = [0, 1, 2, 5, 11, 15, 16, 17, 19, 23, 26, 27, 28, 29, 33, 34, 36, 37, 39, 40, 42]
columns_to_keep = [train_sessions_engineered.columns[ind] for ind in mrmr] 
print(columns_to_keep)

for col in train_sessions_engineered.columns:
    if not col in columns_to_keep:
        train_sessions_engineered = train_sessions_engineered.drop(col)

train_df = train_sessions_engineered.join(train_purchases,train_sessions_engineered.session_id == train_purchases.session_id,"inner" ).cache()
for col in ['session_id','date']:
    train_df = train_df.drop(col)

train_df.take(1)

[0,1,2,4,10,14,15,16,18] ->

#### One Hot Encoding

dic_ohe = {0:4,1:4,4:50,5:50,6:50}
broadcast_d = sc.broadcast(dic_ohe)
# one hot encode all the rows
def ohe(x,bd):
    temp = []
    dic = bd.value
    for i in range(0,len(x)):
        if i in dic:
            new_r = [0]*dic[i]
            if x[i] != -1:
                new_r[int(x[i])] = 1
            temp += new_r
        else:
            temp.append(x[i])
    return tuple(temp)

temp = train_df.rdd.map(lambda x : ohe(x,broadcast_d)).cache()

len(temp.take(1)[0])

#### Create training df

training_data, testing_data = temp.randomSplit([0.9, 0.1])
sampled_training_data, _ = training_data.randomSplit([0.2,0.8])

#training_data.take(1)

# KNN

def pred_knn(data,input_row):
    distance_set = data.map(lambda x : (np.sum(np.abs(np.subtract(x[:-1], input_row))),x[-1])).filter(lambda x : x[0] < 20)
    sorted_set = distance_set.sortBy(lambda x: x[0], ascending=False)
    return(sorted_set.take(100))

true_label = testing_data.take(1)[0][-1]

pred_test = pred_knn(sampled_training_data,testing_data.take(1)[0][:-1])

def score(top_100,label):
    score = 0
    for count, elem in enumerate(top_100):
        if elem[1] == label:
            score = (100-count)/100
            return score
    return score

score(pred_test,true_label)

total_score = []
tests = testing_data.take(10)
for row in tests:
    total_score.append(score(pred_knn(sampled_training_data,row[:-1]),row[-1]))

np.mean(total_score)

pred = np.array([10785., 15542., 27202., 12518., 24465., 15844., 15593., 19372.,
       21675., 17715.,   485., 10111., 24887., 14379.,  1641.,  5272.,
       10304.,   409., 10382.,  8807., 11249.,  3679., 17164., 10724.,
        5724., 21040.,  3334., 17585.,  9538.,  5570.,  1592., 12872.,
       20008., 14652.,  4990., 22736., 11313., 21160.,  4747., 14539.,
       22840., 23293., 10945., 15121., 23657.,  5369.,  4899.,  8434.,
       20724.,  4296., 20604., 17864., 25632., 16029., 12912., 19320.,
       10278., 26258.,  1513., 17005., 19757.,  3720.,  9344., 16441.,
       21001., 24888., 22811., 27086., 27989., 14988.,  5282., 13746.,
       13537., 10295.,  6157., 21819., 18258., 22331.,  3966.,  8017.,
        6762., 20258., 27937., 25308., 27403., 25081., 15188., 14160.,
       14451., 26373., 20083.,  1960., 13955., 16368., 17461.,  3319.,
        3184., 27819., 10976., 10605.])
rank = np.arange(100,0,-1)
id_ = [1238]*100

p = [list(id_),list(pred),list(rank)]
inp = list(map(list, zip(*p)))
print(inp)#spark.createDataFrame(inp,]).show()

list(map(list, zip(*p)))
