==> PreProcessing/3_NaiveBayes.ipynb <==
import os 
import warnings
import time
import numpy as np
import scipy.stats
import sys
import sklearn
import sklearn.datasets

from pyspark.sql import SparkSession
warnings.filterwarnings('ignore')
import pandas as pd

# launch this cell if you have issues on windows with py4j (think about updating your PATH)

os.environ['PYSPARK_DRIVER_PYTHON_OPTS']= "notebook"
os.environ['PYSPARK_DRIVER_PYTHON'] = sys.executable
os.environ['PYSPARK_PYTHON'] = sys.executable

# starts a spark session from notebook

os.environ['PYSPARK_SUBMIT_ARGS'] ="--conf spark.driver.memory=4g  pyspark-shell"


spark = SparkSession \
    .builder \
    .master("local[*]") \
    .appName("feature_selection") \
    .getOrCreate()

sc=spark.sparkContext

#train_sessions_engineered = spark.read.csv('../Data/session_engineered_features.csv',header=False,
#                                          inferSchema=True)
train_sessions_engineered = spark.read.load('../Data/session_engineered_features.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true').cache()


train_sessions_engineered = spark.read.load('../Data/TEMPtrain_sessions.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

train_purchases = spark.read.load('../Data/train_purchases.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

train_purchases.printSchema()

mrmr = [0,1, 2, 5, 11, 15, 16, 17, 19, 23, 26, 27, 28, 29, 33, 34, 36, 37, 39, 40, 42]
columns_to_keep = [train_sessions_engineered.columns[ind] for ind in mrmr] 
print(columns_to_keep)

for col in train_sessions_engineered.columns:
    if not col in columns_to_keep:
        train_sessions_engineered = train_sessions_engineered.drop(col)

# BINING 
from pyspark.ml.feature import Bucketizer
to_bin = ['session_time','std_time']#,'1', '5', '8', '9', '10', '11', '15', '16', '18', '19', '21', '22', '24']
for col in to_bin:
    print(col)
    column_values = train_sessions_engineered.select(col).collect()
    splits_list = [0]
    for i in range(10):
        p = 10 * (i+1)
        splits_list.append(np.percentile(column_values, p))
    splits_list.append(float('Inf'))
    bucketizer = Bucketizer(splits=list(np.unique(splits_list)),inputCol=col, outputCol=col+"_binned")
    train_sessions_engineered = bucketizer.setHandleInvalid("keep").transform(train_sessions_engineered)
    train_sessions_engineered = train_sessions_engineered.drop(col)
    train_sessions_engineered = train_sessions_engineered.withColumnRenamed(col+"_binned",col)

# join the dataframes on the session ids and drop useless columns
train_df = train_sessions_engineered.join(train_purchases,train_sessions_engineered.session_id == train_purchases.session_id,"inner" ).cache()
for col in ['session_id','date']:
    train_df = train_df.drop(col)

train_df.take(1) 

train_df.write.option("header",True).csv('/PROJ/Data/data_processed_for_modelV2.csv')

# NAIVE BAYES

from pyspark.mllib.regression import LabeledPoint
from pyspark.mllib.classification import NaiveBayes, NaiveBayesModel
from pyspark.mllib.util import MLUtils

train_df = spark.read.load('/PROJ/Data/data_processed_for_modelV2.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true').cache()

def if_p(row):
    row_bis = []
    for count,elem in enumerate(row):
        if elem == -1:
            row_bis.append(0)
        else:
            row_bis.append(elem)
    return row_bis

train_df = train_df.rdd.map(if_p)

def labelData(data):
    return data.map(lambda x: LabeledPoint(int(x[-1]), x[:-1]))

training_data, testing_data = labelData(train_df).randomSplit([0.9, 0.1])

model = NaiveBayes.train(training_data)

import numpy
from pyspark.mllib.linalg import _convert_to_vector
from pyspark import RDD

# define custom predict function 
def custom_predict(model_bayes, x):
    labels_ = model_bayes.labels
    pi = model_bayes.pi
    theta = model_bayes.theta
    if isinstance(x, RDD):
        return x.map(lambda v: custom_predict(model_bayes,v))
    x = _convert_to_vector(x)
    x = pi + x.dot(theta.transpose())
    indices = np.argsort(x)[-100:]
    top100 = labels_[indices]
    return top100 #labels_[numpy.argmax(pi + x.dot(theta.transpose()))]


testing_data.map(lambda p: (model.predict(p.features), p.label)).take(20)

model.predict(testing_data.take(4)[3].features)

# TOP 100 with bayes pred

top100 = custom_predict(model,testing_data.take(4)[3].features)
print(top100)

results_top100 = testing_data.map(lambda p: (p.label in custom_predict(model,p.features), p.label))

results_top100.take(10)

results_top100.map(lambda x : (x[0],(1))).reduceByKey(lambda x,y : np.add(x,y)).take(2)

#### VALIDATION SET SCORE : 

def score(top_100,label):
    if label in top_100:
        return (np.where(np.array(top_100) == label)[0][0] + 1)/100
    else:
        return 0

testing_data.map(lambda p: (score(custom_predict(model,p.features),p.label))).mean()

# Competition Prediction

test_df = spark.read.load('/PROJ/Data/test_engineered_features.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')

train_sessions_engineered = spark.read.load('../Data/TEMPtrain_sessions.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true')


test_df.take(1)

mrmr = [0,1, 2, 5, 11, 15, 16, 17, 19, 23, 26, 27, 28, 29, 33, 34, 36, 37, 39, 40, 42]
columns_to_keep = [test_df.columns[ind] for ind in mrmr] 
print(columns_to_keep)

for col in test_df.columns:
    if not col in columns_to_keep:
        test_df = test_df.drop(col)

# BINING 
from pyspark.ml.feature import Bucketizer
to_bin = ['session_time','std_time']#,'1', '5', '8', '9', '10', '11', '15', '16', '18', '19', '21', '22', '24']
for col in to_bin:
    print(col)
    column_values = train_sessions_engineered.select(col).collect()
    splits_list = [0]
    for i in range(10):
        p = 10 * (i+1)
        splits_list.append(np.percentile(column_values, p))
    splits_list.append(float('Inf'))
    bucketizer = Bucketizer(splits=list(np.unique(splits_list)),inputCol=col, outputCol=col+"_binned")
    test_df = bucketizer.setHandleInvalid("keep").transform(test_df)
    test_df = test_df.drop(col)
    test_df = test_df.withColumnRenamed(col+"_binned",col)

test_row = test_df.take(1)[0][1:]
print(test_row)

test_pred = custom_predict(model,test_row)

test_pred # where last item = top1 prediction

array([26428., 24243.,  8535., 26565., 16073., 15542.,  3679., 22719.,
        2967., 24870., 21890.,  2952., 18868., 22886., 10372., 26585.,
        4899., 19525., 19372., 18112., 20093.,  7464., 20303., 26249.,
       22863.,  4171.,  9406.,  5005., 12621., 23578., 21578., 23367.,
       11923.,  5021., 26836., 11275.,    74., 21616., 19145., 13310.,
       21272.,  5433., 15593.,  7750., 12975., 26451., 18489., 23719.,
       11804., 20612.,  6196., 19528., 19402., 28109.,  8188., 26853.,
        7902.,  7096.,   955., 18289., 20449., 19348., 22524.,  1148.,
       15424.,  3438., 10338., 17571.,  4611., 16766.,  2188., 12555.,
        4364., 23450., 22175., 13987., 22209.,  6736., 13656., 21343.,
        6762., 20257., 27393., 18509., 24736., 15249., 16929.,   340.,
         436., 14648.,  4254., 21215., 16846.,  8807.,   403.,  5463.,
        8577.,  4193., 27852., 23789.])

array([10785., 15542., 27202., 12518., 24465., 15844., 15593., 19372.,
       21675., 17715.,   485., 10111., 24887., 14379.,  1641.,  5272.,
       10304.,   409., 10382.,  8807., 11249.,  3679., 17164., 10724.,
        5724., 21040.,  3334., 17585.,  9538.,  5570.,  1592., 12872.,
       20008., 14652.,  4990., 22736., 11313., 21160.,  4747., 14539.,
       22840., 23293., 10945., 15121., 23657.,  5369.,  4899.,  8434.,
       20724.,  4296., 20604., 17864., 25632., 16029., 12912., 19320.,
       10278., 26258.,  1513., 17005., 19757.,  3720.,  9344., 16441.,
       21001., 24888., 22811., 27086., 27989., 14988.,  5282., 13746.,
       13537., 10295.,  6157., 21819., 18258., 22331.,  3966.,  8017.,
        6762., 20258., 27937., 25308., 27403., 25081., 15188., 14160.,
       14451., 26373., 20083.,  1960., 13955., 16368., 17461.,  3319.,
        3184., 27819., 10976., 10605.])

rank = [i for i in range(100,0,-1)]
df = []
for count, input_row in enumerate(test_df.collect()):
    if not count+1 % 5000 :
        print(count)
    session_id = int(input_row[0])
    pred = custom_predict(model,input_row[1:])
    df += [[session_id, int(pred[i]), rank[i]] for i in range(100)]
output_df = spark.createDataFrame([[session_id, int(pred[i]), rank[i]] for i in range(100)])

output_df = spark.createDataFrame(df,['session_id','item_id','rank'])

output_df.coalesce(1).write.option("header",True).csv('/PROJ/Data/leaderboard_prediction.csv')

output_df.count()

output_df = spark.read.load('/PROJ/Data/leaderboard_prediction.csv', 
                          format='com.databricks.spark.csv', 
                          header='true', 
                          inferSchema='true').cache()

output_df.take(10)
